refactor : 중복되는 로직 abstract클래스로 이동
