+++ src/main/java/calculator/Expression.java
@@ -2,5 +2,5 @@ package calculator;
 
 public interface Expression {
 
-    public NumberExpression calc();
+    NumberExpression calc();
 }
