+++ src/main/java/calculator/PlusExpression.java
@@ -1,19 +1,13 @@
 package calculator;
 
-public class PlusExpression implements Expression {
-
-    private Expression leftExpression;
-    private Expression rightExpression;
+public class PlusExpression extends OperationExpression {
 
     public PlusExpression(Expression leftExpression, Expression rightExpression) {
-        this.leftExpression = leftExpression;
-        this.rightExpression = rightExpression;
+        super(leftExpression, rightExpression);
     }
 
     @Override
-    public NumberExpression calc() {
-        NumberExpression left = leftExpression.calc();
-        NumberExpression right = rightExpression.calc();
+    protected NumberExpression execute(NumberExpression left, NumberExpression right) {
         return left.sum(right);
     }
 }
