+++ src/main/java/calculator/SubtractExpression.java
@@ -1,20 +1,13 @@
 package calculator;
 
-public class SubtractExpression implements Expression {
-
-    private Expression leftExpression;
-    private Expression rightExpression;
+public class SubtractExpression extends OperationExpression {
 
     public SubtractExpression(Expression leftExpression, Expression rightExpression) {
-        this.leftExpression = leftExpression;
-        this.rightExpression = rightExpression;
+        super(leftExpression, rightExpression);
     }
 
     @Override
-    public NumberExpression calc() {
-        NumberExpression left = leftExpression.calc();
-        NumberExpression right = rightExpression.calc();
-
+    protected NumberExpression execute(NumberExpression left, NumberExpression right) {
         return left.subtract(right);
     }
 }
